Move the existing routing code to `routing/road`, and create a new folder (`routing/pt`) for PT routing
