+++ src/fleetctrl/rideparcelpooling/immediate/insertion.py
@@ -4,7 +4,7 @@ from src.fleetctrl.FleetControlBase import FleetControlBase
 from src.fleetctrl.planning.VehiclePlan import BoardingPlanStop, VehiclePlan
 from src.fleetctrl.planning.PlanRequest import PlanRequest
 from src.simulation.Vehicles import SimulationVehicle
-from src.routing.NetworkBase import NetworkBase
+from src.routing.road.NetworkBase import NetworkBase
 from src.fleetctrl.pooling.immediate.insertion import simple_insert
 from src.misc.globals import *
 import numpy as np
